% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chocolate_rating_documentation.R
\docType{data}
\name{chocolate_rating}
\alias{chocolate_rating}
\title{Chocolate Ratings Dataset}
\format{
A data frame with 2530 rows and 10 columns:
\describe{
\item{\strong{ref}}{Reference ID. The highest REF numbers were the last entries made.}
\item{\strong{company_manufacturer}}{Name of the company and the manufacturer (manufacturer inside parentheses)}
\item{\strong{company_location}}{The location where the company is based}
\item{\strong{review_date}}{The year the bar was reviewed}
\item{\strong{country_of_bean_origin}}{Country where the cocoa beans originated}
\item{\strong{specific_bean_origin_or_bar_name}}{Specific bean or bar name}
\item{\strong{cocoa_percent}}{Cocoa percent (\% chocolate)}
\item{\strong{ingredients}}{Ingredients (e.g., B = Beans, S = Sugar, S* = Sweetener other than white cane or beet sugar, C = Cocoa Butter, V = Vanilla, L = Lecithin, Sa = Salt)}
\item{\strong{most_memorable_characteristics}}{A summary review of the most memorable characteristics of the bar}
\item{\strong{rating}}{Rating between 1–5}

}
}
\source{
\url{https://flavorsofcacao.com/flavor.html}
}
\usage{
chocolate_rating
}
\description{
This dataset contains information about chocolate ratings from various manufacturers.
}
\keyword{datasets}
